+++ src/parser.js
@@ -34,6 +34,9 @@ export const extract = (code, keyWord) => {
   if (target) {
     const match = target.match(/(?:\()(.*)(?:\))/);
     if (match) {
+      if (match[1] === "") {
+        return []
+      }
       return match[1].split(",").map((item) => item.replace(/\s*/g, ""));
     }
   }